tests: Move all expecter helpers to libvmi
This patch merely moves all the expecter related code to libvmi. It also
moved one virt-launcher helper under libvmi to avoid import cycle with
utils.go.
