+++ tests/libvmi/BUILD.bazel
@@ -4,19 +4,29 @@ go_library(
     name = "go_default_library",
     srcs = [
         "cloudinit.go",
+        "expecter.go",
         "factory.go",
         "network.go",
+        "pod.go",
         "storage.go",
         "vmi.go",
     ],
     importpath = "kubevirt.io/kubevirt/tests/libvmi",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/controller:go_default_library",
+        "//pkg/util/net/dns:go_default_library",
         "//staging/src/kubevirt.io/client-go/api/v1:go_default_library",
+        "//staging/src/kubevirt.io/client-go/kubecli:go_default_library",
+        "//staging/src/kubevirt.io/client-go/log:go_default_library",
         "//tests/containerdisk:go_default_library",
+        "//vendor/github.com/google/goexpect:go_default_library",
+        "//vendor/github.com/onsi/ginkgo:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/resource:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/util/rand:go_default_library",
+        "//vendor/k8s.io/utils/net:go_default_library",
+        "@org_golang_google_grpc//codes:go_default_library",
     ],
 )
