+++ tests/vmi_monitoring_test.go
@@ -30,6 +30,7 @@ import (
 	v1 "kubevirt.io/client-go/api/v1"
 	"kubevirt.io/client-go/kubecli"
 	"kubevirt.io/kubevirt/tests"
+	"kubevirt.io/kubevirt/tests/libvmi"
 )
 
 var _ = Describe("Health Monitoring", func() {
@@ -63,7 +64,7 @@ var _ = Describe("Health Monitoring", func() {
 			launchVMI(vmi)
 
 			By("Expecting the VirtualMachineInstance console")
-			expecter, err := tests.LoggedInAlpineExpecter(vmi)
+			expecter, err := libvmi.LoggedInAlpineExpecter(vmi)
 			Expect(err).ToNot(HaveOccurred())
 			defer expecter.Close()
 
@@ -72,7 +73,7 @@ var _ = Describe("Health Monitoring", func() {
 				&expect.BSnd{S: "watchdog -t 2000ms -T 4000ms /dev/watchdog && sleep 5 && killall -9 watchdog\n"},
 				&expect.BExp{R: "\\#"},
 				&expect.BSnd{S: "echo $?\n"},
-				&expect.BExp{R: tests.RetValue("0")},
+				&expect.BExp{R: libvmi.RetValue("0")},
 			}, 250*time.Second)
 			Expect(err).ToNot(HaveOccurred())
