+++ tests/vmi_multiqueue_test.go
@@ -35,6 +35,7 @@ import (
 	"kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/api"
 	"kubevirt.io/kubevirt/tests"
 	cd "kubevirt.io/kubevirt/tests/containerdisk"
+	"kubevirt.io/kubevirt/tests/libvmi"
 )
 
 var _ = Describe("MultiQueue", func() {
@@ -73,7 +74,7 @@ var _ = Describe("MultiQueue", func() {
 			tests.WaitForSuccessfulVMIStartWithTimeout(vmi, 360)
 
 			By("Checking if we can login")
-			e, err := tests.LoggedInFedoraExpecter(vmi)
+			e, err := libvmi.LoggedInFedoraExpecter(vmi)
 			Expect(err).ToNot(HaveOccurred())
 			e.Close()
 		})
